将 GameFunctions 独立出来
